解决 php 8.1 兼容问题 (#206)
* 测试8.1

* 添加 ReturnTypeWillChange 注解

* 捕获框架初始化错误

* 解决 trait 静态属性被弃用问题

* 调整测试

* 修复代码样式

* 补齐测试环境

* 调整测试

* 测试

* 测试

* 修正

* 测试

* 切换版本

* 修正

* 修正

* 修正

* 修正

* 测试

* 测试

* 切换版本

* 对可能卡住的项目添加超时

* 对可能的发生错误的步骤总是继续执行

* 更新版本

* 恢复全部测试

* 调整

* 修复 windows

* 禁用 windows

* 测试 redis 编译

* 移除 windows php 8.1 测试

* 移除弃用警告忽略

* 优化代码

* 恢复调试输出

+++ src/Components/pgsql/tests/config/config.php
@@ -62,7 +62,7 @@ return [
             'resource'    => [
                 'dbClass'     => 'PdoPgsqlDriver',
                 'host'        => imiGetEnv('PGSQL_SERVER_HOST', '127.0.0.1'),
-                'port'        => imiGetEnv('PGSQL_SERVER_PORT', 5432),
+                'port'        => (int) imiGetEnv('PGSQL_SERVER_PORT', 5432),
                 'username'    => imiGetEnv('PGSQL_SERVER_USERNAME', 'root'),
                 'password'    => imiGetEnv('PGSQL_SERVER_PASSWORD', 'root'),
                 'database'    => 'db_imi_test',
@@ -84,7 +84,7 @@ return [
             'resource'    => [
                 'dbClass'     => 'PdoPgsqlDriver',
                 'host'        => imiGetEnv('PGSQL_SERVER_HOST', '127.0.0.1'),
-                'port'        => imiGetEnv('PGSQL_SERVER_PORT', 5432),
+                'port'        => (int) imiGetEnv('PGSQL_SERVER_PORT', 5432),
                 'username'    => imiGetEnv('PGSQL_SERVER_USERNAME', 'root'),
                 'password'    => imiGetEnv('PGSQL_SERVER_PASSWORD', 'root'),
                 'database'    => 'db_imi_test',
@@ -105,7 +105,7 @@ return [
             'resource'    => [
                 'dbClass'     => 'SwoolePgsqlDriver',
                 'host'        => imiGetEnv('PGSQL_SERVER_HOST', '127.0.0.1'),
-                'port'        => imiGetEnv('PGSQL_SERVER_PORT', 5432),
+                'port'        => (int) imiGetEnv('PGSQL_SERVER_PORT', 5432),
                 'username'    => imiGetEnv('PGSQL_SERVER_USERNAME', 'root'),
                 'password'    => imiGetEnv('PGSQL_SERVER_PASSWORD', 'root'),
                 'database'    => 'db_imi_test',
@@ -120,7 +120,7 @@ return [
             'maindb' => [
                 'dbClass'     => 'PdoPgsqlDriver',
                 'host'        => imiGetEnv('PGSQL_SERVER_HOST', '127.0.0.1'),
-                'port'        => imiGetEnv('PGSQL_SERVER_PORT', 5432),
+                'port'        => (int) imiGetEnv('PGSQL_SERVER_PORT', 5432),
                 'username'    => imiGetEnv('PGSQL_SERVER_USERNAME', 'root'),
                 'password'    => imiGetEnv('PGSQL_SERVER_PASSWORD', 'root'),
                 'database'    => 'db_imi_test',
